# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/rfq": {
      "post": {
        "operationId": "RfqController_createRFQ",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRfqDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create RFQ (Buyer only)",
        "tags": [
          "RFQ"
        ]
      }
    },
    "/api/v1/rfq/{id}/publish": {
      "put": {
        "operationId": "RfqController_publishRFQ",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishRfqDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Publish RFQ (Buyer only)",
        "tags": [
          "RFQ"
        ]
      }
    },
    "/api/v1/rfq/me": {
      "get": {
        "operationId": "RfqController_getMyRFQs",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get my RFQs (Buyer)",
        "tags": [
          "RFQ"
        ]
      }
    },
    "/api/v1/rfq/{id}": {
      "get": {
        "operationId": "RfqController_getRFQById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get RFQ by ID",
        "tags": [
          "RFQ"
        ]
      }
    },
    "/api/v1/rfq/open": {
      "get": {
        "operationId": "RfqController_getOpenRFQs",
        "parameters": [
          {
            "name": "category",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "grade",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get open RFQs (Seller view)",
        "tags": [
          "RFQ"
        ]
      }
    },
    "/api/v1/rfq/{id}/close": {
      "put": {
        "operationId": "RfqController_closeRFQ",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Close/Cancel RFQ (Buyer)",
        "tags": [
          "RFQ"
        ]
      }
    },
    "/api/v1/quotes": {
      "post": {
        "operationId": "QuotesController_submitQuote",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit quote for RFQ (Seller only)",
        "tags": [
          "Quotes"
        ]
      }
    },
    "/api/v1/quotes/me": {
      "get": {
        "operationId": "QuotesController_getMyQuotes",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get my quotes (Seller)",
        "tags": [
          "Quotes"
        ]
      }
    },
    "/api/v1/quotes/rfq/{rfqId}": {
      "get": {
        "operationId": "QuotesController_getQuotesForRFQ",
        "parameters": [
          {
            "name": "rfqId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get quotes for RFQ (Buyer)",
        "tags": [
          "Quotes"
        ]
      }
    },
    "/api/v1/quotes/{id}": {
      "get": {
        "operationId": "QuotesController_getQuoteById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get quote by ID",
        "tags": [
          "Quotes"
        ]
      }
    },
    "/api/v1/quotes/{id}/accept": {
      "put": {
        "operationId": "QuotesController_acceptQuote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptQuoteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Accept quote (Buyer only)",
        "tags": [
          "Quotes"
        ]
      }
    },
    "/api/v1/quotes/{id}/withdraw": {
      "put": {
        "operationId": "QuotesController_withdrawQuote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Withdraw quote (Seller)",
        "tags": [
          "Quotes"
        ]
      }
    },
...
  },
  "info": {
    "title": "B2B Marketplace API",
    "description": "Backend API for B2B Marketplace Platform",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateRfqDto": {
        "type": "object",
        "properties": {}
      },
      "PublishRfqDto": {
        "type": "object",
        "properties": {}
      },
      "CreateQuoteDto": {
        "type": "object",
        "properties": {}
      },
      "AcceptQuoteDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}